gcp_idp_project/requirements.txt
surya-ocr==0.17.1
transformers
torch>=2.7.0,<3.0.0
accelerate
Pillow
python-docx
pandas
numpy
sentencepiece
pypdfium2
protobuf
requests
opencv-python-headless
